Move datastructure packages under felix.
