+++ go/felix/endpoint/endpoint_policy_resolver.go
@@ -16,8 +16,8 @@ package endpoint
 
 import (
 	log "github.com/Sirupsen/logrus"
-	"github.com/projectcalico/felix/go/datastructures/multidict"
-	"github.com/projectcalico/felix/go/datastructures/set"
+	"github.com/projectcalico/felix/go/felix/multidict"
+	"github.com/projectcalico/felix/go/felix/set"
 	"github.com/projectcalico/libcalico-go/lib/backend/api"
 	"github.com/projectcalico/libcalico-go/lib/backend/model"
 )
