+++ go/felix/felix.go
@@ -20,11 +20,11 @@ import (
 	log "github.com/Sirupsen/logrus"
 	"github.com/docopt/docopt-go"
 	pb "github.com/gogo/protobuf/proto"
-	"github.com/projectcalico/felix/go/datastructures/ip"
 	"github.com/projectcalico/felix/go/felix/buildinfo"
 	"github.com/projectcalico/felix/go/felix/calc"
 	"github.com/projectcalico/felix/go/felix/config"
 	_ "github.com/projectcalico/felix/go/felix/config"
+	"github.com/projectcalico/felix/go/felix/ip"
 	"github.com/projectcalico/felix/go/felix/logutils"
 	"github.com/projectcalico/felix/go/felix/proto"
 	"github.com/projectcalico/felix/go/felix/statusrep"
