+++ go/felix/set/set.go
